@@ -19,37 +19,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.4.4",
-    "@angular/compiler": "2.4.4",
-    "@angular/core": "2.4.4",
-    "@angular/forms": "2.4.4",
-    "@angular/http": "2.4.4",
-    "@angular/material": "2.0.0-alpha.9-3",
-    "@angular/platform-browser": "2.4.4",
-    "@angular/platform-browser-dynamic": "2.4.4",
-    "@angular/router": "3.4.5",
-    "angular-calendar": "0.6.2",
-    "angular2-materialize": "~6.6.0",
-    "angular2-recaptcha": "^0.3.7",
-    "core-js": "^2.4.1",
-    "dragula": "^3.7.2",
+    "@angular/animations": "^9.0.0-rc.7",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^9.0.0-rc.7",
+    "@angular/compiler": "^9.0.0-rc.7",
+    "@angular/core": "^9.0.0-rc.7",
+    "@angular/forms": "^9.0.0-rc.7",
+    "@angular/material": "^9.0.0-rc.6",
+    "@angular/platform-browser": "^9.0.0-rc.7",
+    "@angular/platform-browser-dynamic": "^9.0.0-rc.7",
+    "@angular/router": "^9.0.0-rc.7",
+    "@fullcalendar/core": "^4.3.1",
+    "angular-calendar": "^0.28.2",
+    "chart.js": "^2.9.3",
+    "core-js": "^2.6.11",
+    "date-fns": "^2.8.1",
     "file-saver": "^1.3.3",
-    "jquery": "^2.2.4",
+    "jquery": "^3.4.1",
+    "kurento-utils": "^6.13.0",
     "materialize-css": "~0.97.8",
-    "ng2-dragula": "~1.2.1",
-    "ng2-file-upload": "1.2.0",
-    "primeng": "~1.1.4",
+    "ngx-materialize": "^6.1.3",
+    "openvidu-browser": "^2.11.0",
+    "primeng": "^9.0.0-rc.2",
     "quill": "~1.1.10",
-    "rxjs": "^5.0.1",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.0.0-rc.0",
     "time-ago-pipe": "~1.1.1",
     "ts-helpers": "^1.1.1",
-    "zone.js": "^0.6.23",
-    "openvidu-browser": "1.8.0"
+    "tslib": "^1.10.0",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
+    "@angular-devkit/build-angular": "^0.803.21",
+    "@angular/cli": "^9.0.0-rc.7",
+    "@angular/compiler-cli": "^9.0.0-rc.7",
+    "@types/core-js": "^2.5.2",
     "@types/jasmine": "2.5.41",
-    "@angular/cli": "1.4.3",
-    "@angular/compiler-cli": "^2.4.0",
     "codelyzer": "~0.0.26",
     "jasmine-core": "2.4.1",
     "jasmine-expect": "^3.0.1",
@@ -64,6 +69,6 @@
     "protractor": "4.0.5",
     "ts-node": "1.2.1",
     "tslint": "3.13.0",
-    "typescript": "2.6.1"
+    "typescript": "^3.6.4"
   }
 }
