@@ -2,18 +2,7 @@ import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from '../app.component';
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule }      from '@angular/core';
 import { FormsModule }   from '@angular/forms';
-import { HttpModule }    from '@angular/http';
-
-import { routing } from '../app.routing';
-
-import { MaterializeModule }   from 'angular2-materialize';
-import { FileSelectDirective } from 'ng2-file-upload';
-import { FileDropDirective }   from 'ng2-file-upload';
-
-import { MaterialModule } from '@angular/material';
-
 import { NavbarComponent }        from '../components/navbar/navbar.component';
 import { FooterComponent }        from '../components/footer/footer.component';
 import { LoginModalComponent }    from '../components/login-modal/login-modal.component';
@@ -29,23 +18,10 @@ import { FileUploaderComponent }  from '../components/file-uploader/file-uploade
 import { StreamComponent }        from '../components/video-session/stream.component';
 import { ChatLineComponent }      from '../components/chat-line/chat-line.component';
 
-import { AuthenticationService }  from '../services/authentication.service';
-import { CourseService }          from '../services/course.service';
-import { SessionService }         from '../services/session.service';
-import { ForumService }           from '../services/forum.service';
-import { FileService }            from '../services/file.service';
-import { CourseDetailsModalDataService }  from '../services/course-details-modal-data.service';
-import { LoginModalService }      from '../services/login-modal.service';
-import { UploaderModalService }   from '../services/uploader-modal.service';
-import { UserService }            from '../services/user.service';
-import { AnimationService }       from '../services/animation.service';
-
-import { CalendarModule }         from 'angular-calendar';
 import { CalendarComponent }      from '../components/calendar/calendar.component';
 import { TimeAgoPipe }            from 'time-ago-pipe';
-import { DragulaModule }          from 'ng2-dragula/ng2-dragula';
-import { EditorModule }           from 'primeng/components/editor/editor';
-import { ReCaptchaModule }        from 'angular2-recaptcha';
+import {HttpClientModule} from '@angular/common/http';
+import {MaterializeModule} from 'ngx-materialize';
 
 
 describe('App: Ejemplo', () => {
@@ -54,14 +30,9 @@ describe('App: Ejemplo', () => {
       imports: [
         BrowserModule,
         FormsModule,
-        HttpModule,
+        HttpClientModule,
         MaterializeModule,
-        MaterialModule.forRoot(),
-        RouterTestingModule,
-        CalendarModule.forRoot(),
-        DragulaModule,
-        EditorModule,
-        ReCaptchaModule,
+        RouterTestingModule
       ],
       declarations: [
         AppComponent,
@@ -77,8 +48,6 @@ describe('App: Ejemplo', () => {
         FileGroupComponent,
         CalendarComponent,
         TimeAgoPipe,
-        FileSelectDirective,
-        FileDropDirective,
         VideoSessionComponent,
         FileUploaderComponent,
         StreamComponent,
