@@ -2,5 +2,5 @@ export const Constants = Object.freeze({
   PHONE_MAX_WIDTH: 500,
   TOAST_SHOW_TIME: 5000,
   PASS_REGEX: '^((?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,20})$',
-  FILE_SIZE_LIMIT: 5*1024*1024 //5MB
+  FILE_SIZE_LIMIT: 5 * 1024 * 1024 // 5MB
 });
